requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
pydub==0.25.1
requests==2.31.0
urllib3==1.26.15
midiutil==1.2.1
python-dotenv==1.0.0
PyJWT==2.8.0
python-docx==0.8.11
python-pptx==0.6.21
PyPDF2==3.0.1
reportlab==4.0.4
Werkzeug==2.3.6
